route: move HTTP response logic to shared package
